init commit:
This tutorial is modified tutorial created by By Christian Lundkvist and Sam Mayo This is a tutorial intended to cover the very basics of the libsnark software library for creating zk-SNARKs. We will demonstrate how to formulate zk-SNARK circuits, create proofs and verify the proofs. The difference between my and their tutorial is that my tutorial serializes verification key and proof to file (aka sends file to Verifier) and then Verifier deserializes vk and proof and verify the SNARK. Here we create proof for knowledge of root of: x^3 + x + 5 == 35 The main focus of this tutorial is in section 2: libsnark components: 2. Gadgets Steps are commented in libsnark-tutorial/src/test-gadget.cpp